fix(apps/medusa): missing type, restructure directories
MDS-60
