+++ apps/medusa/src/api/store/customers/me/rewards/route.ts
@@ -4,13 +4,15 @@ import {
 } from '@medusajs/framework';
 
 import { LOYALTY_MODULE } from '../../../../../modules/loyalty';
+import LoyaltyModuleService from '../../../../../modules/loyalty/service';
 
 export const GET = async (
   req: AuthenticatedMedusaRequest,
   res: MedusaResponse
 ) => {
   const query = req.scope.resolve('query');
-  const loyaltyModuleService = req.scope.resolve(LOYALTY_MODULE);
+  const loyaltyModuleService: LoyaltyModuleService =
+    req.scope.resolve(LOYALTY_MODULE);
 
   const loyaltyPoints = await loyaltyModuleService.getPoints(
     req.auth_context.actor_id
