+++ apps/medusa/src/workflows/loyalty/apply-loyalty-on-cart/steps/get-cart-loyalty-promo-amount.ts
@@ -2,8 +2,8 @@ import { CustomerDTO, PromotionDTO } from '@medusajs/framework/types';
 import { MedusaError } from '@medusajs/framework/utils';
 import { StepResponse, createStep } from '@medusajs/framework/workflows-sdk';
 
-import { LOYALTY_MODULE } from '../../../modules/loyalty';
-import LoyaltyModuleService from '../../../modules/loyalty/service';
+import { LOYALTY_MODULE } from '../../../../modules/loyalty';
+import LoyaltyModuleService from '../../../../modules/loyalty/service';
 
 export type GetCartLoyaltyPromoAmountStepInput = {
   cart: {
