+++ apps/medusa/src/workflows/loyalty/update-reward/index.ts
@@ -5,7 +5,7 @@ import {
 } from '@medusajs/framework/workflows-sdk';
 import { useQueryGraphStep } from '@medusajs/medusa/core-flows';
 
-import { linkProductToRewardStep } from './steps/link-product-to-reward';
+import { linkProductToRewardStep } from '../steps/link-product-to-reward';
 import { updateRewardStep } from './steps/update-reward';
 
 type UpdateRewardWorkflowInput = {
